Refactor file locations and remove garbage
