+++ examplePlugin/.babelrc
