+++ examplePlugin/.editorconfig
