+++ examplePlugin/.eslintrc
