+++ examplePlugin/.gitignore
