+++ examplePlugin/package.json
