+++ examplePlugin/src/js/main.js
